test(utils): move crypto tests into an independent dir
