+++ packages/ckb-sdk-utils/__tests__/crypto/bech32.fixtures.json
